requirements.txt
langchain
langgraph
tavily-python
fastapi
faiss-cpu
uvicorn
